Document type

+++ construction/views.py
@@ -178,6 +178,7 @@ class DashboardGraph(APIView):
                 print(disbursement_schedule(loan_id))
                 queryset = DisbursementSchedule.objects.filter(loan_id =loan_id).order_by('review_months')
                 serializer = DisbursementScheduleSerializer(queryset, many=True)
+                return Response(disbursement_schedule(loan_id))
             case 'construction_status_graph':
                 queryset = ConstructionStatus.objects.filter(loan_id =loan_id).order_by('review_months')
                 serializer = ConstructionStatusSerializer(queryset, many=True)
@@ -397,6 +398,8 @@ class CreateRetrieveUpdateLoan(APIView):
         project_id = request.data.get('project')   
         try:
             project_type = Project.objects.get(pk=project_id).project_type
+            if Loan.objects.filter(project_id=project_id).exists():
+                return Response({"detail": "This project is already associated with a loan."}, status=status.HTTP_409_CONFLICT)
         except Project.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
         input_json['loantype'] = project_type
@@ -409,6 +412,7 @@ class CreateRetrieveUpdateLoan(APIView):
         
         if serializer.is_valid():
             loan = serializer.save()
+
             document_details = DocumentDetail.objects.filter(document_type_id__in = list(document_type_obj))
             document_details_list = list()
             for detail in document_details:
@@ -622,11 +626,25 @@ class CreateUpdateDrawRequest(APIView):
             else:
                 val = draw_req_obj.order_by('-draw_request').values_list('draw_request',flat=True)
                 draw_request = val[0] +1
+                # Check the status of the previous draw from the DrawTracking table
+                last_draw = DrawTracking.objects.get(loan_id=loan_id,draw_request=val[0])
+                # If the last draw exists and its status is "Pending" or "In Review", prevent new draw creation
+                if last_draw and last_draw.draw_status in ["Pending", "In Review"]:
+                    return Response(
+                         {"error":"The previous draw is still 'Pending' or 'In Review'. Cannot create a new draw."},
+                         status=status.HTTP_400_BAD_REQUEST
+                    )
+
+                
                 for obj in budget_master_obj:
                     budget_amount = obj[1]  
                     released_amount_previous = DrawRequest.objects.filter(budget_master_id=obj).values_list('released_amount',flat=True)        
                     released_amount_list = list(released_amount_previous)
-                    released_amount=released_amount_list[0]
+                    if released_amount_list:
+                        released_amount=released_amount_list[0]
+                    else:
+                        released_amount = 0
+                    
                     new_instance = DrawRequest(
                         budget_master_id=obj[0],
                         draw_request = draw_request,                    
@@ -732,11 +750,12 @@ class CreateUpdateDrawRequest(APIView):
             draw_request_obj = DrawRequest.objects.filter(
                 budget_master_id__in= budget_master_id,
                 draw_request = draw_request
-            )
+            ).order_by('budget_master__uses_type','budget_master__uses')
+            
         else:
             draw_request_obj = DrawRequest.objects.filter(
                 budget_master_id__in= budget_master_id
-            )
+            ).order_by('budget_master__uses_type','budget_master__uses')
         serializers = DrawRequestSerializer(draw_request_obj,many=True)
         return Response(serializers.data, status=status.HTTP_200_OK)
     
@@ -756,12 +775,15 @@ class RetrieveDeleteUpdateDrawTracking(APIView):
     def delete(self, request, id):
         try:
             draw_tracking_obj = DrawTracking.objects.get(pk=id)
-            draw_request = draw_tracking_obj.draw_request
-            loan_id = draw_tracking_obj.loan
-            draw_tracking_obj.delete()
-            budget_master_ids = BudgetMaster.objects.filter(loan_id=loan_id).values_list('id',flat=True)
-            DrawRequest.objects.filter(draw_request=draw_request, budget_master_id__in = budget_master_ids).delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
+            if draw_tracking_obj.draw_status in ['Pending','In Review']: 
+                draw_request = draw_tracking_obj.draw_request
+                loan_id = draw_tracking_obj.loan
+                draw_tracking_obj.delete()
+                budget_master_ids = BudgetMaster.objects.filter(loan_id=loan_id).values_list('id',flat=True)
+                DrawRequest.objects.filter(draw_request=draw_request, budget_master_id__in = budget_master_ids).delete()
+                return Response(status=status.HTTP_204_NO_CONTENT)
+            else:
+                return Response({"error":"Only draw requests with 'Pending' or 'In Review' status can be deleted."},status=status.HTTP_403_FORBIDDEN)
         except DrawTracking.DoesNotExist:
             return Response({"error": "DrawTracking object not found."}, status=status.HTTP_404_NOT_FOUND)
         except Exception as e:
